Configuration moved to WIFI
Enabled WIFI configuration for WIFI, as well as google script, OpenW API and longitude/lattitude.

This makes it easier to gift the calender to a friend :)

It will spin a WIFI portal up (using the Autoconnect library) both in case it does not find WIFI, and for 5 min every time you hard-boot (not deep-sleep).